Finished functionality to register and subscribe for events and organization, missing to add more data and styling 2

+++ app/src/main/java/com/example/bradj/eventit/LoginActivity.java
@@ -335,7 +335,7 @@ public class LoginActivity extends AppCompatActivity implements LoaderCallbacks<
                     public void onResponse(Call<User> call, Response<User> response) {
                         if(response.isSuccessful()) {
                             User user=response.body();
-                            loginUtil.createUserSession(getApplicationContext(),response.body().getUserId(), user);
+                            loginUtil.createUserSession(getApplicationContext(),(int)response.body().getUserId(), user);
                             Intent intent = new Intent(getApplicationContext(), MainActivity.class);
                             startActivity(intent);
                             finish();
